refactor(coderd): tidy up NATS cluster mTLS after review
Drop the redundant ClusterCAKeycache interface in favor of cryptokeys.SigningKeycache, remove the unused CryptoKey.DecodeString method, merge the NATS TLS integration test into the internal test file, and note the CA-cache boot refactor as a TODO.
